requirements.txt
aiocoap==0.4.7
cefevent==0.5.6
confluent-kafka==2.3.0
fastavro==1.9.4
paho_mqtt==2.0.0
pika==1.3.2
pysyslogclient==0.1.1
python-dotenv==1.0.1
requests==2.31.0
